requirements.txt
# Core RL Libraries
gymnasium>=0.29.0
stable-baselines3>=2.0.0
ray[rllib]>=2.8.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0

# Visualization and Analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
streamlit>=1.25.0

# Configuration and Utilities
pyyaml>=6.0
hydra-core>=1.3.0
wandb>=0.15.0
tensorboard>=2.13.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Additional RL Utilities
tianshou>=0.5.0
pfrl>=0.3.0
